Substitution and weakening, separated at last.
Simultaneous substitutions require the definitions of substitution
and weakening to be mutually recursive, which is needlessly complicated
and gets worse with dependent types. Explicit substitutions are supposed
to do both substitution and weakening at once, but I still don't
understand how.

With embedded weakening, weakening can be implemented first and
independently of substitution. Then, as a second step, substitution can
be implemented in terms of weakening. I think it's much cleaner that
way...