CoContainers, who would have thought?
Containers have a really concise definition for fmap.
Can we find a similar family of types for which cofmap
can be defined as easily?

The initial intuition was that if both containers and
the functor (U -> a) have left-composition for fmap, then
cocontainers should have right-composition for fmap, just
like (a -> U). The definition of cocontainers quickly followed.