DepCont positions aren't dependent: renaming...
DepCont elements are dependent, but not positions.
It's possible to combine dependencies, but there is a combinatorial
number of combinations! in order to keep sanity, we need sensible names.

I think it would be wiser to have a prefix for containers with
dependent positions, not for containers without; but that's the
convention I'm currently using, so I'll stick to it.

DepCont shall henceforth be known as FlatDepCont.