fmap generalizations
noticing that fmap and cofmap each operated on one of the three type
parameters, I conjectured that a third kind of map function, smap, could
operate on the shape parameter. I was right.

furthermore, mapping functions exist for each non-empty subset of the
three parameters (again, causing a combinatorial explosion). however,
these functions don't exist for all types of containers.

if a container variable B is dependent on a container variable A, then a
mapping function modifying A will only be defined if the mapping
function also modifies B. for this reason, fmap (which only changes
elements) and dismap (which changes everything) are always defined.