turning adjunctions into records
categories are records, functors are records, and now so are
adjunctions. also, adjunctions are parameterized over categories now,
not functors. the functors are kept as fields.

this makes it obvious that the "composition-preserves-adjunction"
property from last time is actually a form of composition! This explains
why the property is so useful.

I had to rewrite almost everything in order to fit the code into this
new straightjacket, but I think it's worth it. unfortunately the code
became so concise that it wasn't clear anymore whether a particular
adjunction composition was indeed relating the functors we thought it
did. the solution: a kind of type annotation which uses the type-checker
to verify that it is so.