a version of (Examples.Nat) using (Fin 2) to tag constructors
Defining a datatype using the technique illustrated in Conor McBride's
Ornaments paper requires choosing a type with two elements to tag the
two constructors (zero) and (suc). Conor uses the set {zz, ss}, which
has the advantage of being readable; this corresponds to our NatCon
in (Example.Nat).

Our version of datatype definitions, however, requires this tag type to
be totally ordered, requiring the user to prove that finite sets such as
{zz, ss} can be totally ordered. This is of course completely trival
since the set is finite, but Agda doesn't know this and the fact must be
proven for each such set.

This commit simplifies the situation by proving that for any (n), the
tag type (Fin n) can be totally ordered. Unfortunately, using instances
of (Fin n) for constructor tags decreases the readability of the code
drastically, so it is not yet clear which of the two approaches, that of
(Example.Nat) and that of (Example.Fin.Nat), is preferable.